requirements.txt
# general modules
Django==4.2.7
djangorestframework


# packagess
python-decouple
Pillow
coreapi
coreschema


# third party modules
django-filter
Markdown
drf-yasg
drf-yasg[validation]
djangorestframework_simplejwt==5.3.1



# deployment module


# email third party modules
django-mail-templated
